+++ .storybook/main.js
@@ -1,7 +1,12 @@
 const path = require('path');
 module.exports = {
   "stories": ["../src/**/*.stories.mdx", "../src/**/*.stories.@(js|jsx|ts|tsx)"],
-  "addons": ["@storybook/addon-links", "@storybook/addon-essentials", "@storybook/addon-interactions"],
+  "addons": [
+    "@storybook/addon-links",
+    "@storybook/addon-essentials",
+    "@storybook/addon-interactions",
+    "@storybook/addon-webpack5-compiler-babel"
+  ],
   "framework": {
     name: "@storybook/react-webpack5",
     options: {}
@@ -23,7 +28,8 @@ module.exports = {
     // Return the altered config
     return config;
   },
-  docs: {
-    autodocs: true
+  docs: {},
+  typescript: {
+    reactDocgen: "react-docgen-typescript"
   }
 };