+++ src/components/FhiButton/__tests__/FhiButton.spec.tsx
@@ -1,10 +1,9 @@
-import renderer from 'react-test-renderer';
+import { render } from '@testing-library/react';
 import Button from '../FhiButton';
 
 describe('FhiButton', () => {
   test('snapshot test', () => {
-    const button = renderer.create(<Button onClick={() => {}}>test</Button>);
-    const tree = button.toJSON();
-    expect(tree).toMatchSnapshot();
+    const { container } = render(<Button onClick={() => {}}>test</Button>);
+    expect(container).toMatchSnapshot();
   });
 });