+++ src/components/FhiDetails/__tests__/FhiDetails.spec.tsx
@@ -1,9 +1,9 @@
-import renderer from 'react-test-renderer';
+import { render } from '@testing-library/react';
 import Details from '../FhiDetails';
 
 describe('FhiDetails', () => {
   test('snapshot test', () => {
-    const details = renderer.create(
+    const { container } = render(
       <Details
         className="test"
         items={[
@@ -14,7 +14,6 @@ describe('FhiDetails', () => {
         ]}
       />,
     );
-    const tree = details.toJSON();
-    expect(tree).toMatchSnapshot();
+    expect(container).toMatchSnapshot();
   });
 });