+++ src/components/FhiSpinner/FhiSpinner.stories.tsx
@@ -1,5 +1,3 @@
-/* eslint-disable import/no-extraneous-dependencies */
-/* eslint-disable react/jsx-props-no-spreading */
 import type { StoryObj, Meta } from '@storybook/react';
 
 import Spinner from './FhiSpinner';