+++ src/components/FhiToast/__tests__/FhiToast.spec.tsx
@@ -1,10 +1,9 @@
-import renderer from 'react-test-renderer';
+import { render } from '@testing-library/react';
 import Toast from '../FhiToast';
 
 describe('FhiToast', () => {
   test('snapshot test', () => {
-    const toast = renderer.create(<Toast text="This is a toast!" removeToast={() => {}} delay={0} />);
-    const tree = toast.toJSON();
-    expect(tree).toMatchSnapshot();
+    const { container } = render(<Toast text="This is a toast!" removeToast={() => {}} delay={0} />);
+    expect(container).toMatchSnapshot();
   });
 });